test(swift): fix tests for int/double differenciation

+++ test/renderers/swift/ast.test.ts
@@ -167,7 +167,7 @@ describe("statementNodeToDeclNode", () => {
           variableDecl({}, SwiftAst.Keyword.let, [
             patternBinding(
               identifierPattern("constantValue"),
-              identifierType("Int")
+              identifierType("Double")
             ),
           ]),
           variableDecl({}, SwiftAst.Keyword.let, [
